# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  http.ErrorResponse:
    properties:
      error:
        example: invalid credentials
        type: string
    type: object
...
  http.DestinationViewCount:
    properties:
      last_updated_at:
        example: "2025-11-12T03:35:49Z"
        type: string
      total_views:
        example: 1287
        format: int64
        type: integer
      unique_ips:
        example: 342
        type: integer
      unique_users:
        example: 12
        type: integer
    type: object
  http.DestinationViewStatsResponse:
    properties:
      city:
        example: Bangkok
        type: string
      country:
        example: Thailand
        type: string
      destination_id:
        example: e3ff134a-bbd0-46f4-9586-2414627acafd
        type: string
      name:
        example: Manual Test Destination
        type: string
      views:
        additionalProperties:
          $ref: '#/definitions/http.DestinationViewCount'
        type: object
    type: object
  http.TrendingDestination:
    properties:
      city:
        example: Chiang Mai
        type: string
      country:
        example: Thailand
        type: string
      destination_id:
        example: 8cda9899-6533-4ed2-bf3e-85ee68588b79
        type: string
      name:
        example: Ao Manao - Khao Tanyong National Park
        type: string
      views:
        additionalProperties:
          $ref: '#/definitions/http.DestinationViewCount'
        type: object
    type: object
  http.DestinationTrendingResponse:
    properties:
      destinations:
        items:
          $ref: '#/definitions/http.TrendingDestination'
        type: array
      limit:
        example: 10
        type: integer
      range:
        example: 24h
        type: string
    type: object
  http.DestinationStatsExportRequest:
    properties:
      destination_ids:
        items:
          example: e3ff134a-bbd0-46f4-9586-2414627acafd
          type: string
        type: array
    type: object
...
  http.PageMeta:
    properties:
      count:
        example: 1
        type: integer
      total:
        example: 57
        type: integer
      limit:
        example: 20
        type: integer
      offset:
        example: 0
        type: integer
    type: object
  http.Destination:
    properties:
      category:
        example: Nature
        type: string
      city:
        example: New York
        type: string
      country:
        example: USA
        type: string
      created_at:
        example: "2024-07-10T12:00:00Z"
        type: string
      deleted_at:
        type: string
      description:
        example: Iconic urban park spanning 843 acres.
        type: string
      contact:
        example: +1-212-310-6600
        type: string
      opening_time:
        example: "08:00"
        type: string
      closing_time:
        example: "22:00"
        type: string
      gallery:
        items:
          $ref: '#/definitions/http.DestinationMedia'
        type: array
      average_rating:
        example: 4.3
        format: double
        type: number
      review_count:
        example: 128
        type: integer
      hero_image_url:
        example: https://cdn.fitcity/destinations/central-park.jpg
        type: string
      id:
        example: 1da5f07d-7dcf-4aea-b948-7c9daa0a5db7
        type: string
      latitude:
        example: 40.785091
        format: double
        type: number
      longitude:
        example: -73.968285
        format: double
        type: number
      name:
        example: Central Park
        type: string
      slug:
        example: central-park
        type: string
      status:
        enum:
        - draft
        - published
        - archived
        type: string
      updated_at:
        example: "2024-07-11T08:00:00Z"
        type: string
      updated_by:
        example: 9fd13fd2-63c5-4f29-a210-4a1a8e285f74
        type: string
      version:
        example: 3
        type: integer
    type: object
  http.DestinationListResponse:
    properties:
      destinations:
        items:
          $ref: '#/definitions/http.Destination'
        type: array
      meta:
        $ref: '#/definitions/http.PageMeta'
    type: object
...
  http.DestinationMedia:
    properties:
      url:
        description: Publicly accessible image URL or CDN path.
        example: https://cdn.fitcity/destinations/central-park/gallery-1.jpg
        type: string
      caption:
        description: Short description rendered under the image.
        example: Summer blooms near the lake
        type: string
      ordering:
        description: Zero-based ordering for display.
        example: 1
        type: integer
    required:
    - url
    type: object
...
info:
  contact: {}
  description: Authentication and profile services for the Fit City platform.
  title: Fit City API
  version: "1.0"
paths:
...
  /destinations:
    get:
      description: List published destinations visible to end users.
      parameters:
      - description: Maximum number of records to return
        in: query
        name: limit
        type: integer
      - description: Number of records to skip
        in: query
        name: offset
        type: integer
      - description: Full text search across name, city, country, category, or description
        in: query
        name: query
        type: string
      - description: Comma-separated list of categories to include
        in: query
        name: categories
        type: string
      - description: Minimum average rating (0-5)
        format: double
        in: query
        name: min_rating
        type: number
      - description: Maximum average rating (0-5)
        format: double
        in: query
        name: max_rating
        type: number
      - description: Country filter
        in: query
        name: country
        type: string
      - description: City filter
        in: query
        name: city
        type: string
      - description: Latitude for nearby search
        in: query
        name: lat
        type: number
        format: double
      - description: Longitude for nearby search
        in: query
        name: lng
        type: number
        format: double
      - description: Max distance in kilometers for nearby search
        in: query
        name: max_distance_km
        type: number
        format: double
      - description: Additional Sort (similarity, distance)
        in: query
        name: sort
        type: string
      - description: Sort order (rating_desc, rating_asc, alpha_asc, alpha_desc, updated_at_desc)
        in: query
        name: sort
        type: strings
      - description: Sort order (rating_desc, rating_asc, alpha_asc, alpha_desc, updated_at_desc)
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.DestinationListResponse'
      summary: List published destinations
      tags:
      - Destinations
...
  /destinations/{identifier}/views:
    get:
      description: Return aggregated view counts for a published destination using its UUID or slug.
      parameters:
      - description: Destination UUID or slug
        in: path
        name: identifier
        required: true
        type: string
      - description: Comma-separated view ranges to include (1h,6h,12h,24h,7d,30d,all). Defaults to all ranges when omitted.
        in: query
        name: range
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.DestinationViewStatsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Get destination view stats
      tags:
      - Destinations
  /destinations/trending:
    get:
      description: List the most-viewed destinations for a given time window.
      parameters:
      - description: View range (1h,6h,12h,24h,7d,30d,all). Defaults to all-time when omitted.
        in: query
        name: range
        type: string
      - description: Maximum number of destinations to return (default 10).
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.DestinationTrendingResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: List trending destinations
      tags:
      - Destinations
...
  /admin/destination-stats/views:
    get:
      description: Return destination view metrics for admins with a forced Elasticsearch refresh.
      parameters:
      - description: Destination UUID
        in: query
        name: destination_id
        required: true
        type: string
      - description: Optional range filter (1h,6h,12h,24h,7d,30d,all). Defaults to all-time.
        in: query
        name: range
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.DestinationViewStatsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Admin destination view stats
      tags:
      - Admin Destinations
  /admin/destination-stats/export:
    post:
      consumes:
      - application/json
      description: Export destination popularity metrics as CSV for the requested destinations. When no body is provided, all published destinations are included.
      parameters:
      - description: Optional list of destination IDs to include
        in: body
        name: data
        schema:
          $ref: '#/definitions/http.DestinationStatsExportRequest'
      produces:
      - text/csv
      responses:
        "200":
          description: CSV export
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Export destination popularity
      tags:
      - Admin Destinations
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"